requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pinecone-client==5.0.1
sentence-transformers==3.1.1
rank-bm25==0.2.2
PyPDF2==3.0.1
python-docx==1.1.2
transformers==4.44.2
torch==2.4.1
numpy==1.26.4
pandas==2.2.3
python-dotenv==1.0.1
langchain==0.3.0
langchain-community==0.3.0
langchain-huggingface==0.1.0